+++ src/main/kotlin/io/github/dlachouette/teamcity/github/web/WebhookInfo.kt
@@ -51,7 +51,7 @@ object WebhookEvents {
     val RECOMMENDED: List<String> = listOf(
         "pull_request",
         "pull_request_review",
-        "issue_comment",
+        "pull_request_review_comment",
         "check_run",
         "push",
         "check_suite",